+++ src/main/java/com/instaclustr/ingenuity/techtest/RandomNumberGeneratorUtil.java
@@ -13,16 +13,26 @@ import java.util.concurrent.ArrayBlockingQueue;
  * @author  Sureshraja
  * @version 1.0
  * @see {@link NumberGenerator } & {@link NumberProcessor}
+ *
+ * Design Explanation:
+ * =====================
+ *
+ * Created the Util class to have all the supporting functions of this application inside this class.
+ *
+ * In getRandomNumber(), Random calss from Util package is used to generate the Random numbers in the range.
+ *
+ * Multisets class from Google's Guava used to get the most common number used in the queue. Also the Stats class is used to compute the average.
+ *
  */
 
 public class RandomNumberGeneratorUtil {
 
-    public static int getRandomNumber(int min, int max) {
+    public static byte getRandomNumber(int min, int max) {
         Random random = new Random();
-        return random.ints(min, (max + 1)).findFirst().getAsInt();
+        return (byte)random.ints(min, (max + 1)).findFirst().getAsInt();
     }
 
-    public  static int getMostCommonNumber(ArrayBlockingQueue<Integer> sharedQueue){
+    public  static int getMostCommonNumber(ArrayBlockingQueue<Byte> sharedQueue){
        return Multisets.copyHighestCountFirst(ImmutableMultiset.copyOf(sharedQueue))
                 .iterator().next();
     }
@@ -36,7 +46,7 @@ public class RandomNumberGeneratorUtil {
         return (String.format("%" + length + "s", "").replace(" ", String.valueOf(car)) + str).substring(str.length(), length + str.length());
     }
 
-    public static void printResults(ArrayBlockingQueue<Integer> sharedQueue, int minRandomValue,
+    public static void printResults(ArrayBlockingQueue<Byte> sharedQueue, int minRandomValue,
                              int maxRandomValue, Integer mostCommon) {
         System.out.format("%s | Min = %s | Max = %s | Avg = %s | Most Common = %s \n",
                 LPad(Arrays.toString(sharedQueue.toArray()), 100, ' '),
